+++ src/rp.jl
@@ -5,16 +5,12 @@
     rp!(X::Matrix, weights::Weight; kwargs...)
 Make a random projection of X-data.
 * `X` : X-data (n, p).
-* `weights` : Weights (n) of the observations. 
-    Must be of type `Weight` (see e.g. function `mweight`).
+* `weights` : Weights (n) of the observations. Must be of type `Weight` (see e.g. function `mweight`).
 Keyword arguments:
 * `nlv` : Nb. dimensions on which `X` is projected.
-* `meth` : Method of random projection. Possible
-    values are: `:gauss`, `:li`. See the respective 
-    functions `rpmatgauss` and `rpmatli` for their 
-    keyword arguments.
-* `scal` : Boolean. If `true`, each column of `X` 
-    is scaled by its uncorrected standard deviation.
+* `meth` : Method of random projection. Possible values are: `:gauss`, `:li`. See the respective 
+    functions `rpmatgauss` and `rpmatli` for their keyword arguments.
+* `scal` : Boolean. If `true`, each column of `X` is scaled by its uncorrected standard deviation.
 
 ## Examples
 ```julia